refactor(platform): T2 SPEC-011 — mover logout/refresh_token a sus carpetas
git mv (4 archivos) desde auth/login/ a sus carpetas correctas:
- auth/login/auth_logout_use_case.py → auth/logout/
- auth/login/auth_refresh_token_use_case.py → auth/refresh_token/
- auth/login/auth_refresh_token_orchestrator_use_case.py → auth/refresh_token/
- auth/login/auth_refresh_token_external_use_case.py → auth/refresh_token/

4 imports actualizados:
- controller/business/auth_controller.py (2 imports)
- auth/refresh_token/auth_refresh_token_orchestrator_use_case.py (2 imports internos)

Respeta el patrón existente "una sub-carpeta por UC" usado en el resto de
auth/ (create_company, create_user_internal, delete_user_internal, etc.).

Smoke test: app bootea con 120 routes; ningún ref viejo a auth.login.auth_logout
o auth.login.auth_refresh queda en src/.

Ref: Goluti/01 - Platform/Specs/SPEC-011-platform-dead-code-cleanup.md
